% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_logodds_associations}
\alias{df_logodds_associations}
\title{Odds Ratios of Blood Biomarkers with Incident Type 2 Diabetes}
\format{A data frame (tibble) with 228 rows and 5 columns:
\describe{
  \item{name}{Biomarker abbreviation.}
  \item{study}{Short description of the association type, here the cohort
  name.}
  \item{beta}{Log odds for incident type 2 diabetes.}
  \item{se}{Standard error.}
  \item{pvalue}{P-value.}
  \item{n}{Sample size.}
}}
\source{
These data are taken from the Supplementary material of
A. V. Ahola-Olli et al. (2019).
\url{https://www.biorxiv.org/content/early/2019/01/08/513648}
}
\usage{
df_logodds_associations
}
\description{
A data frame containing log odds ratios of the Nightingale blood biomarkers
with risk for future type 2 diabetes.
}
\details{
"Values are odds ratios (95\% confidence intervals) per 1-SD log-transformed
metabolite concentration. Odds ratios were adjusted for sex, baseline age,
BMI, and fasting glucose. YFS, Cardiovascular risk in Young Finns Study;
NFBC, Northern Finland Birth Cohort."
The values are shown in Figure S3 of A. V. Ahola-Olli et al. (2019):
\url{https://www.biorxiv.org/content/early/2019/01/08/513648}
}
\keyword{datasets}
