% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_NG_biomarker_metadata}
\alias{df_NG_biomarker_metadata}
\title{Information and Grouping for Nightingale's Blood Biomarkers}
\format{A data frame (tibble) with 252 rows and 8 columns:
\describe{
  \item{abbreviation}{Biomarker abbreviation, i.e. the one delivered with the
  xlsx data format.}
  \item{machine_readable_name}{Biomarker machine readable name, i.e. the one
  delivered with the csv data format.}
  \item{name}{Biomarker name.}
  \item{alternative_names}{Alternative biomarker names.}
  \item{description}{Biomarker description.}
  \item{group}{The group the biomarker belongs to.}
  \item{subgroup}{The subgroup the biomarker belongs to.}
  \item{unit}{Unit is deprecated. Use the units delivered with biomarkers.}
}}
\source{
Nightingale Health Ltd. \url{https://nightingalehealth.com/}
}
\usage{
df_NG_biomarker_metadata
}
\description{
A data frame with information on the Nightingale Health Ltd. NMR-quantified
blood biomarkers.
}
\keyword{datasets}
