% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforestplot-package.R
\docType{package}
\name{ggforestplot-package}
\alias{ggforestplot}
\alias{ggforestplot-package}
\title{ggforestplot: Forestplot Visualizations.}
\description{
ggforestplot is a package for plotting measures of effect, e.g. linear
associations or log and hazard ratios, with their confidence interval in a
forestplot layout. It's main plotting function is \code{forestplot()}.
}
\author{
\strong{Maintainer}: Maria Kalimeri \email{maria.kalimeri@nightingalehealth.com} (0000-0002-6569-7553)

Authors:
\itemize{
  \item Ilari Scheinin \email{ilari.scheinin@nightingalehealth.com} (0000-0002-4696-9066)
  \item Juuso Parkkinen \email{juuso.parkkinen@nightingalehealth.com} (0000-0002-7818-5901)
  \item Emmi Tikkanen \email{emmi.tikkanen@nightingalehealth.com} (0000-0003-1634-4134)
  \item Peter Würtz \email{peter.wurtz@nightingalehealth.com} (0000-0002-5832-0221)
  \item Antti Kangas \email{antti.kangas@nightingalehealth.com}
}

Other contributors:
\itemize{
  \item Nightingale Health Ltd. \email{contact@nightingalehealth.com} [copyright holder]
}

}
\keyword{internal}
