% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_colours.R
\name{ng_colour}
\alias{ng_colour}
\alias{display_ng_colours}
\title{Nightingale's colours}
\usage{
ng_colour(...)

display_ng_colours()
}
\arguments{
\item{...}{Character names of Nightingale's colours.}
}
\description{
\code{ng_colour()} returns Nightingale's colours' hex codes.
\code{display_ng_colours()} displays the avaliable colours, with their names
and hex codes, in a vertical layout.
}
\examples{
# Display Nightingale's colours
display_ng_colours()

# Request for the hex code of colour 'dark pesto'
ng_colour("dark pesto")
}
\author{
Ilari Scheinin
}
