% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_colours.R
\name{ng_palette_d}
\alias{ng_palette_d}
\alias{ng_palette_c}
\alias{display_ng_palettes}
\title{Nightingale's colour palettes}
\usage{
ng_palette_d(name = "all", reverse = FALSE)

ng_palette_c(name = "magma", reverse = FALSE, ...)

display_ng_palettes()
}
\arguments{
\item{name}{Character name of the Nightingale (or viridis) colour palette.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to
\code{\link[grDevices:colorRampPalette]{colorRampPalette()}}}
}
\description{
\code{ng_palette_d()} and \code{ng_palette_c()} (respectively for discrete
and continuous palettes) return functions that take an integer argument (the
required number of colours) and return a character vector of colours' hex
codes.
In addition, the functions also recognize the
\code{\link[viridisLite:viridis]{viridis}} palettes: "magma" (or "A"),
"inferno" ("B"), "plasma" ("C"), "viridis" ("D"), or "cividis" ("D").
}
\examples{
#' # Display Nightingale's colour palettes
display_ng_palettes()

# Get 4 colours along the spectrum of the nwr palette
ng_palette_d("nwr")(4)

# Notice that the discrete palette "light", cannot return more than 5 colours
ng_palette_d("light")(6)

}
\author{
Ilari Scheinin
}
