% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_forest.R
\name{theme_forest}
\alias{theme_forest}
\title{Forestplot Theme}
\usage{
theme_forest(base_size = 13, base_line_size = base_size/22,
  base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A custom theme used in \code{\link{forestplot}} that builts upon
\link[ggplot2]{theme_minimal}.
}
\seealso{
\code{\link{forestplot}}, \code{\link{geom_effect}}, \code{\link{geom_stripes}}
}
\author{
Maria Kalimeri
}
