% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_grouping_all_NG_biomarkers}
\alias{df_grouping_all_NG_biomarkers}
\title{Example of Biomarker Grouping Data Frames for Nightingale Biomarkers}
\format{A data frame (tibble) with 2 rows and 2 columns:
\describe{
  \item{version}{Biomarker platform version.}
  \item{layout}{A data frame (tibble) defining layout:
  \itemize{
  \item \code{machine_readable_name} Biomarker machine readable name, i.e. the one
  delivered with the csv data format.
  \item \code{group_custom} A character indication group titles.
  \item \code{column} An integer indicating the column number in a
  layout, see \code{\link{plot_all_NG_biomarkers}}.
  \item \code{page} An integer indicating the page number in a
  layout, see \code{\link{plot_all_NG_biomarkers}}.
  }
}
}}
\usage{
df_grouping_all_NG_biomarkers
}
\description{
A data frame containing example custom groupings for the Nightingale Health
biomarkers in a machine readable name format. There are two custom groupings
available for 2016 and 2020 biomarker platforms. These groupings are used
with \code{\link{plot_all_NG_biomarkers}}.
}
\seealso{
Data frame \code{\link{df_NG_biomarker_metadata}} with information
on the Nightingale Health Ltd. NMR-quantified blood biomarkers
}
\keyword{datasets}
